bugfix: properly retrieve the openshift version
oc_version_output was being set to the string "get_version.stdout"
rather than the value of the variable "get_version.stdout". This was
causing the values of "openshift_provision_oc_version" and
"openshift_provision_openshift_version" to be set to the string
"get_version.stdout" which is not correct.

Fixes #17