use resource fqdn when running patch and delete
If there are multiple resources with the same `Kind` but are part of
different API groups, you have to provide the API group and the Kind
in the format `kind.apigroup` to ensure the correct resource is
modified.

For example, in OpenShift 4.x, there are `configs.operator.openshift.io`
and `configs.imageregistry.operator.openshift.io` resources. If you
try to run `oc patch configs cluster...`, the API call would not know for
sure if you were trying to patch the image registry config, or the
operator config and can fail.

To prevent issues, you can provide the full resource kind in your
command, for example to patch the image registry, you should use
`oc patch configs.imageregistry.operator.openshift.io cluster...`

Fixes #19